[Support] Factor out unused op pruning from LLHD Mem2Reg; NFC
Factor out the tracking and pruning of unused, side effect free ops from
the Mem2Reg pass into a new `UnusedOpPruner` utility. This just wraps a
set of ops and provides helper functions to mark ops, values, and value
ranges as to be pruned if they end up being unused at a later point.

This helps passes clean up after themselves if they remove ops that may
render other parts of the IR unused. I'm working on an LLHD pass that
will benefit from this utility.
